+++ bypy/gvar.py
@@ -15,7 +15,12 @@ from . import const
 try:
 	SystemLanguageCode, SystemEncoding = locale.getdefaultlocale()
 except ValueError as e:
-	if e.args and e.args[0] and e.args[0] == "unknown locale: UTF-8": # macOS 常见
+	# https://coderwall.com/p/-k_93g/mac-os-x-valueerror-unknown-locale-utf-8-in-python
+	# Mac OS X: ValueError: unknown locale: UTF-8 in Python
+	# Proper fix:
+	# export LC_ALL=en_US.UTF-8
+	# export LANG=en_US.UTF-8
+	if e.args and e.args[0] and e.args[0] == "unknown locale: UTF-8":
 		SystemLanguageCode, SystemEncoding = '', 'UTF-8'
 	else:
 		raise